Fix project list command naming
Correctly use "project list" as the command name instead of "org projects".
